Fix all remaining linting warnings and errors for CI/CD checks
