+++ src/__tests__/setup.js
@@ -4,7 +4,7 @@
  */
 
 import "@testing-library/jest-dom";
-import { vi } from "vitest";
+import { vi, beforeEach, afterEach } from "vitest";
 
 // Mock window.matchMedia
 Object.defineProperty(window, "matchMedia", {
